Implementation of LogStore class
Implementing all the functions in the LogStore class with the possibilty to create the logstring on the fly. It is possible to enable/disable this functionality.

The LogStore class will be used to store the logs and organize them in the correct format.
Then, it will be possible to export the logs in a string format or in a dictionary format (see med-material/LoggingManager#13).