Implementing the WriteToCsv class
This class will allow to save the logs in a file.
Each LogToCSV object is instantiated with a file path and a LogStore object.
When calling the WriteAll() function, the logs are exported from the LogStore object as a string and are written in the file at the given file path.
The process is currently executed in a separated thread to prevent the main application from freezing.